Add list predicates to predicates.py
